requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
mysqlclient==2.2.7
python-decouple==3.8
setuptools==69.2.0
sqlparse==0.5.3
wheel==0.43.0
